refactor: Nits on the code & move utils to key-utils
