+++ src/library-authoring/create-unit/CreateUnitModal.tsx
@@ -13,7 +13,7 @@ import messages from './messages';
 import { useCreateLibraryContainer } from '../data/apiHooks';
 import { ToastContext } from '../../generic/toast-context';
 import LoadingButton from '../../generic/loading-button';
-import { ContainerType } from '../utils';
+import { ContainerType } from '../../generic/key-utils';
 
 const CreateUnitModal = () => {
   const intl = useIntl();
